diffnet-paddlepaddle/requirements.txt
argon2-cffi==20.1.0
astor==0.8.1
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
blessings==1.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
conda==4.8.3
conda-package-handling==1.7.0
contextlib2==0.6.0.post1
coverage==5.2.1
cryptography==2.9.2
Cython==0.29.24
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
future==0.18.2
gast==0.4.0
gpustat==0.6.0
hyperopt==0.1.2
idna==2.9
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
json-tricks==3.15.2
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab-server==1.2.0
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
netifaces==0.10.9
networkx==2.5
nni==1.8
notebook==6.1.3
numpy==1.19.1
nvidia-ml-py3==7.352.0
packaging==20.4
paddlepaddle==2.1.3
paddlepaddle-gpu==2.1.3.post110
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pgl==2.1.5
pickleshare==0.7.5
Pillow==7.2.0
pkginfo==1.5.0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.6
protobuf==3.19.1
psutil==5.7.2
ptyprocess==0.6.0
pycosat==0.6.3
pycparser==2.20
Pygments==2.6.1
pymongo==3.11.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
PythonWebHDFS==0.2.3
pyzmq==19.0.2
qtconsole==4.7.6
QtPy==1.9.0
requests==2.23.0
ruamel-yaml==0.15.87
ruamel.yaml.clib==0.2.0
schema==0.7.2
scikit-learn==0.23.2
scipy==1.5.2
Send2Trash==1.5.0
simplejson==3.17.2
six==1.14.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
torch==1.6.0
torchvision==0.7.0
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
widgetsnbextension==3.5.1
